+++ src/packages/excalidraw/CHANGELOG.md
@@ -28,6 +28,7 @@ Please add the latest change on the top under the correct section.
 
 - Rename `appState.elementLocked` to `appState.activeTool.locked` [#4983](https://github.com/excalidraw/excalidraw/pull/4983).
 - Expose [`serializeLibraryAsJSON`](https://github.com/excalidraw/excalidraw/blob/master/src/packages/excalidraw/README.md#serializeLibraryAsJSON) helper that we use when saving Excalidraw Library to a file.
+- Replace `appState.exportWithDarkMode` with `appState.exportTheme`.
 
 ##### BREAKING CHANGE
 
@@ -38,6 +39,10 @@ You will need to pass `activeTool.locked` instead of `elementType` from now onwa
 ##### BREAKING CHANGE
 
 You will need to pass `activeTool` instead of `elementType` from now onwards in `appState`
+
+##### BREAKING CHANGE
+
+You will need to replace usage of `exportWithDarkMode` by `exportTheme`
 
 ### Build
 
@@ -64,6 +69,7 @@ React.createElement(ExcalidrawLib.Excalidraw, opts);
 #### Features
 
 - Support system theme in the `DarkModeToggle` component
+- Add a new system theme to use when exporting SVGs
 
 #### Refactor